Update artifact file size column type to BigInteger in migration and model
- Changed the file_size column type from Integer to BigInteger in the artifacts table migration.
- Updated the Artifact model to reflect the new BigInteger type for file_size, ensuring compatibility with larger file sizes.

+++ lncrawl/migrations/versions/2026_03_19_21_33_54_artifact_file_size_column.py
@@ -22,7 +22,7 @@ depends_on: Union[str, Sequence[str], None] = None
 
 def upgrade() -> None:
     """Upgrade schema."""
-    op.add_column("artifacts", sa.Column("file_size", sa.Integer(), server_default=sa.literal(0), nullable=False))
+    op.add_column("artifacts", sa.Column("file_size", sa.BigInteger(), server_default=sa.literal(0), nullable=False))
 
     conn = op.get_bind()
     executor = TaskManager(15)
